refactor: 프로바이더 구조 리팩토링 및 레이아웃 수정
- ProvidersLayout 제거 및 개별 프로바이더 파일로 분리
- AuthProvider, ProfileProvider, ToastProvider 파일 추가
- Providers 컴포넌트 생성 및 레이아웃에 적용
- 불필요한 코드 정리 및 리팩토링
