+++ src/app/_provider/ReactQuery.provider.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { PropsWithChildren } from 'react';
