+++ src/app/_provider/Toast.provider.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+
+export default function ToastProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+      <ToastContainer />
+    </>
+  );
+}
