+++ src/app/_provider/index.tsx
@@ -1,0 +1,16 @@
+import AuthProvider from './Auth.provider';
+import ProfileProvider from './Profile.provider';
+import ReactQueryProvider from './ReactQuery.provider';
+import ToastProvider from './Toast.provider';
+
+export default function Providers({ children }: { children: React.ReactNode }) {
+  return (
+    <AuthProvider>
+      <ReactQueryProvider>
+        <ProfileProvider>
+          <ToastProvider>{children}</ToastProvider>
+        </ProfileProvider>
+      </ReactQueryProvider>
+    </AuthProvider>
+  );
+}
