+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import Background from '@/app/_containers/Layout';
 import '@app/globals.css';
 import type { Metadata } from 'next';
 import { Roboto } from 'next/font/google';
+import Providers from './_provider';
 
 export const metadata: Metadata = {
   title: 'HomeLog',
@@ -22,7 +23,9 @@ export default function HTMLLayout({
   return (
     <html lang='en'>
       <body className={`${roboto.className} bg-[#DADADA]`}>
-        <Background>{children}</Background>
+        <Providers>
+          <Background>{children}</Background>
+        </Providers>
       </body>
     </html>
   );
