+++ src/contexts/auth.context.tsx
@@ -1,13 +1,5 @@
 'use client';
-import api from '@/app/api';
-import {
-  createContext,
-  PropsWithChildren,
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-} from 'react';
+import { createContext, useContext } from 'react';
 
 type AuthContextType = {
   isLoggedIn: boolean;
@@ -16,7 +8,7 @@ type AuthContextType = {
   signOut: () => void;
 };
 
-const AuthContext = createContext<AuthContextType>({
+export const AuthContext = createContext<AuthContextType>({
   isLoggedIn: false,
   loading: true,
   signIn: () => {},
@@ -24,38 +16,5 @@ const AuthContext = createContext<AuthContextType>({
 });
 
 const useAuth = () => useContext(AuthContext);
-
-export function AuthProvider({ children }: PropsWithChildren<any>) {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [loading, setLoading] = useState(true);
-
-  const signIn = useCallback(() => setIsLoggedIn(true), []);
-
-  const signOut = useCallback(async () => {
-    await api.auth.signOut();
-    setIsLoggedIn(false);
-  }, []);
-
-  useEffect(() => {
-    const checkLoginStatus = async () => {
-      setLoading(true);
-      try {
-        const response = await api.auth.checkSignIn();
-        setIsLoggedIn(response);
-      } catch (error) {
-        setIsLoggedIn(false);
-      }
-      setLoading(false);
-    };
-
-    checkLoginStatus();
-  }, []);
-
-  return (
-    <AuthContext.Provider value={{ isLoggedIn, loading, signIn, signOut }}>
-      {children}
-    </AuthContext.Provider>
-  );
-}
 
 export default useAuth;
