+++ src/contexts/profile.context.tsx
@@ -1,7 +1,5 @@
 'use client';
-import { createContext, PropsWithChildren, useContext, useEffect } from 'react';
-import useAuth from './auth.context';
-import useQueryGetProfile from '@/hooks/profile/useQuery.getProfile';
+import { createContext, useContext } from 'react';
 
 export type Profile = {
   nickname: string;
@@ -10,24 +8,6 @@ export type Profile = {
   homeImageUrl: string;
 };
 
-const ProfileContext = createContext<Profile | null>(null);
+export const ProfileContext = createContext<Profile | null>(null);
 
 export const useProfile = () => useContext(ProfileContext);
-
-export function ProfileProvider({ children }: PropsWithChildren) {
-  const isLoggedIn = useAuth();
-
-  const { data: profile, isFetched: isProfileFetched } = useQueryGetProfile();
-
-  useEffect(() => {
-    if (!isLoggedIn) return;
-    if (!isProfileFetched) {
-      return;
-    }
-  }, [isLoggedIn, isProfileFetched]);
-  return (
-    <ProfileContext.Provider value={profile || null}>
-      {children}
-    </ProfileContext.Provider>
-  );
-}
